Change from address precedences.
Now the -f arg overrides the from address in pmailrc. They are used for the
envelope from address. If there is no from header on stdin then they are also
used to fill this in. If there is a from header on stdin and there is no from
address either on the command line or in the config file then that header is
used for the envelope from address too.