Encode body as quoted printable if not 7bit clean.
Only applies for multi-part mails, for now.

Previously for multipart mails the body is automatically encoded as the
default MIMEText encoding, base64. This made it less readable for readers that
don't support multipart, ie when reading the raw mail.

Now we don't encode the body if it is 7bit clean; we assume it is ascii. If
the body is not 7bit clean (ie, there are 8bit+ chars involved) then we ecnode
it as utf-8 quoted printable.

The way this is done is to set the default encoding for utf-8 text as QP. Then
to detect whether the body is 7bit ascii or something else (presumably utf-8)
we use the magic package.

The other option was to use the string.encode('ascii') function, and if that
failed assume it was utf-8. But that function doesn't work properly in my
opinion. If the passed text has, for example, the copyright character (which
is 8 bits and hence outside the range of ascii) at passes and MIMEText set the
charset to 8bit. Since there is no such thing as 8bit ascii (right?) this is
broken.