Handle friendly names with comma.
For splitting the recipients headers (To, CC, BCC) previously I was naïvely
splitting on comma, this breaks on people with friendly names like

    "Last, First" <first.last@example.com>

Now we use some pythonic regular expression to not worry about commas in
double quotes.

gatt: Also changes the printout when attachments are saved to a directory. Al
slash is added on the end for emphasis.