Generate Message-ID header before sending.
The Message-ID header was previously being generated by whatever
transport element got the sent message first. Since I sending via a ope
SMTP gateway I have to manually put the sent mail in my Maildir. When I
go to reply to a message that I sent myself (when appending to a thread
I was the last to reply to for instance) notmuch was generating its own
message ID to put in In-Reply-To and References. Since this was
different that the Message-ID that the other parties received confusion
abounded (right?).

We now add a Message-ID header via pythons email.Utils.make_msgid().
Note that to reduce information leakage you may want to change this to
use the domain part of the From: address rather than the local FQDN.