Suport sending from a system account.
System wide config file.

No /dev/tty.

Don't worry about permission denied errors when writing to sent folder because
probably it is being sent from a system account and the mailidr from the
system wide config is a user's one.

Other possible solutions to that are to just use a different config file for
the user, our to have the user's one clobber stuff from the system one. Or to
check if the dest addr is going to end up for local delivery anyway.