Allow converting markdown messages to markdown+html.
Converts the message body with pymarkdown (extras,sane_lists) to html and adds
a multipart/alternative part as the first part of a message with both the
original body and the html part in that (in that order). The markup=markdown
parameter is added to the "Content-Type: text/plain" header for the markdown
part so that MUAs that are aware of that can do some prettier rendering.

I would have liked to add the format=flowed paramater too but the
"space-stuffing" "feature" of the RFC is too much of a pain with markdown
(like with lists and tables and code blocks. Maybe I will add that in the
Format head too so it is on by default but can be switched off for more
complex emails.

TODO: Allow a format header that can by default is markdown but can be changed
to plain or html (or maybe markdown,html).
TODO: Turn link URLs that point to attachments (by filename) into cd:...
links, and give attachments Content-ID headers.