Implement format=flowed (rfc3676).
If you set the flowed param (currently not implemented) the text/plain portion
of the message WILL BE MANGLED to conform with the (poorly thought out IMO)
rfc3676. Namely:

* Add the format=flowed param to the Content-Type header.
* Prepend lines starting with spaces or From with a space.
      * The RFC says to prepend unquoted lines that start with > too, but
        we have no way of knowing which lines are quoted vs unquoted.
* Append lines in a paragraph which can have the following line appended for
  reflowing with a space.
      * Actually more simplisticly: append to every line in a block but the
        last one. A block is a sequence of non-zero lines.

The rfc states that line separators should be \r\n but I am sticking with
pythons email.as_string() which seems to use \n. Maybe that is platform
dependant.