html: Sub img link sources for Content-ID.
Substitute img link source attributes for the Content-ID of the attachment
part with the same filename. This should allow better support for displaying
inline images in HTML enable MUAs[1]. The Content-Disposition of the relevant
parts is set to inline too.

For the actual substitution we are using re.sub but sadly we have to do it
line-by-line because appartly pythons regex implementation only treats ^ as
the start of the string, not the start of a line. The pattern isn't complied,
but how many lines are your email bodies anyway? Also, maybe you want a proper
html parser or too do the substition on tha markdown and make sure the line
isn't indented. Meh.

NOTE: Only supports replacing image sources which are basenames of the
attached files. As that is what the Content-Disposition filename param is
reduced to.

TODO: Add all the inline images into a multipart/related part. That means the
the main body will be in a heir of mixed->related->alternative->text though.

[1]: http://stackoverflow.com/a/4697643