Fix/tweak quoted-printible line endings with format=flowed.
Format=flowed (rfc2646) soft line endings (MAY re-flow, ' \n') were getting
encoded as ' =\n\n'. While this quoted-printable encoding ended up with the
decoded text being the same it made more more painful viewing of the encoded
text because it now has a bunch of blank lines in it. Another valid Q-P
encoding of this sequence is '=20\n' which mantains the same number of lines
in the output and this is also what the other (rfc1521) python
quoted-printable module does. So now I am using that when required.

The older module probably hase different output in other cases too but if that
is an issue it is easy enough to change the newer one to support this:

    --- /usr/lib/python2.7/email/quoprimime.py 22:59:16.800590409 +1200
    +++ /usr/lib/python2.7/email/quoprimime.py 22:59:48.484589857 +1200
    @@ -245,7 +245,7 @@
                         encoded_body += encoded_line + prev
                 # Just normal whitespace at end of line
                 else:
    -                encoded_body += encoded_line + prev + '=' + eol
    +                encoded_body += encoded_line + quote(prev)
                 encoded_line = ''
             # Now look at the line we just finished and it has a line ending, we
             # need to add eol to the end of the line.