Add options for markdown, html, flowed, qp.
These options default to false (apart from qp) except when the editor is
launched when they are set true and temporarily add to a 'Format' header. The
goal of this is that we don't want to screw with automatically sent emails but
if a person is sending it assume they are sending the common type of
person-to-person prose email, if not they can change the options.