Fall back to the magic module for attachment mime types.
The python mimetypes module only looks up the filename extension in a table.
So a text attachment without an extension woudl hev been base64ed and set to
application/octet-stream.

Also changed it so compressed types are not set to octet-stream. This was
copied from https://docs.python.org/2/library/email-examples.html but I realy
can't see a reason for it. Now if someone has a handler set up for
application/x-tar (for instance) they will get a chance to handle it without
having to parse it themselves.